refactor: simplify incredible-avs crate
